Use buffered IO for stderr
Previously we called fprintf all over the code and used unbuffered IO
so that we didn't need to add the explicit flush all over the code. Now
that we're only calling fprintf from within the logger we can explicit
flush the stream from there. The motivation for using buffered io is that
some platforms may print character by character when using unbuffred IO.