New commands and flags for integrating checkpoint support into TAP.
This adds the new commands and flags to the TAP protocol for
the integration of checkpoints into the TAP module. Especially,
TAP_CHECKPOINT_START and TAP_CHECKPOINT_END notifies the TAP
client of the start and end of a checkpoint, respectively, and
are used for checkpoint synchronization between the TAP server
and client.